requirements.txt
asgiref==3.3.4
autopep8==1.5.7
dateutils==0.6.12
dj-database-url==0.5.0
Django==3.2.3
django-crispy-forms==1.11.2
django-decouple==2.1
django-heroku==0.3.1
django-rest-framework-braces==0.3.4
djangorestframework==3.12.4
gunicorn==19.9.0
Pillow==8.2.0
psycopg2==2.8.6
pycodestyle==2.7.0
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
